Adding ABC for api class
